Refactor measureText function
